koltech web/research/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
playwright==1.36.0
markdown==3.4.4
jsonschema==4.19.0
robotexclusionrulesparser==1.7.1
aiohttp==3.8.5
tqdm==4.65.0
